+++ apps/console/src/extensions/configs/attribute.ts
@@ -277,7 +277,7 @@ export const attributeConfig: AttributeConfig = {
             showOnlyMandatory: true,
             showPrimaryUserStore: false,
             showReadOnlyAttribute: false,
-            showRegularExpression: false,
+            showRegularExpression: true,
             showSummary: false
         },
         customDialectURI: "urn:scim:wso2:schema",