+++ apps/console/src/features/branding/api/use-get-branding-preference-resolve.ts
@@ -25,7 +25,7 @@ import useRequest, {
     RequestResultInterface
 } from "../../core/hooks/use-request";
 import { OrganizationType } from "../../organizations/constants/organization-constants";
-import { useGetOrganizationType } from "../../organizations/hooks/use-get-organization-type";
+import { useGetCurrentOrganizationType } from "../../organizations/hooks/use-get-organization-type";
 import {
     BrandingPreferenceAPIResponseInterface,
     BrandingPreferenceTypes
@@ -45,7 +45,7 @@ const useGetBrandingPreferenceResolve = <Data = BrandingPreferenceAPIResponseInt
         type: BrandingPreferenceTypes = BrandingPreferenceTypes.ORG,
         locale: string = I18nConstants.DEFAULT_FALLBACK_LANGUAGE
     ): RequestResultInterface<Data, Error> => {
-    const { organizationType } = useGetOrganizationType();
+    const { organizationType } = useGetCurrentOrganizationType();
 
     const tenantDomain: string = organizationType === OrganizationType.SUBORGANIZATION
         ? store.getState()?.organization?.organization?.id