+++ apps/console/src/features/server-configurations/pages/index.ts
@@ -16,7 +16,6 @@
  * under the License.
  */
 
-export * from "./admin-session-advisory-banner-edit-page";
 export * from "./governance-connectors";
 export * from "./connector-edit-page";
 export * from "./connector-listing-page";