+++ modules/i18n/src/translations/si-LK/portals/console.ts
@@ -8759,6 +8759,10 @@ export const console: ConsoleNS = {
                     pageHeading: "පරිපාලක උපදේශන බැනරය",
                     pageSubheading: "පිවිසුම් පිටුවේ පෙන්වීමට පරිපාලක උපදේශන බැනරය වින්‍යාස කරන්න."
                 },
+                manageNotificationSendingInternally: {
+                    title: "අභ්යන්තර දැනුම්දීම යැවීම",
+                    description: "දැනුම්දීම කළමනාකරණය කිරීම අභ්යන්තරව යවයි."
+                },
                 remoteLogPublishing: {
                     title: "දුරස්ථ ලොග් ප්‍රකාශනය",
                     pageTitle: "දුරස්ථ ලොග් ප්‍රකාශනය",
@@ -8907,6 +8911,10 @@ export const console: ConsoleNS = {
                             }
                         }
                     }
+                },
+                server: {
+                    title: "සේවාදායකය",
+                    description: "සේවාදායක සැකසුම් වින්යාස කරන්න."
                 }
             },
             sidePanel: {
@@ -9321,6 +9329,7 @@ export const console: ConsoleNS = {
                         generic: {
                             inputs: {
                                 placeholder: "ඔබගේ {{fieldName}} ඇතුළත් කරන්න",
+                                dropdownPlaceholder: "ඔබගේ {{fieldName}} තෝරන්න",
                                 validations: {
                                     empty: "{{fieldName}} යනු අත්‍යවශ්‍ය ක්ෂේත්‍රයකි",
                                     invalidFormat: "{{FieldName}} නිවැරදි ආකෘතියෙන් නොවේ"